refactor(suite): general fiat service created
