+++ packages/suite/src/services/fiat/index.ts
@@ -1,0 +1,9 @@
+import * as coingecko from './coingecko';
+
+export const {
+    getTickerConfig,
+    fetchCurrentTokenFiatRates,
+    fetchCurrentFiatRates,
+    fetchLastWeekRates,
+    getFiatRatesForTimestamps,
+} = coingecko;
